+++ src/waitForTheElement.js
@@ -5,7 +5,7 @@
  *
  * @private
  *
- * @returns {Element} The matching element from the mutation, otherwise null.
+ * @returns {Element} The matching element from the mutation, otherwise `null`.
  *
  * @param {MutationRecord} mutation The DOM mutation.
  * @param {String}         selector The selector that the element must match.
@@ -49,13 +49,13 @@ function getMatchingElementFromMutation  (mutation, selector)
  * ```
  * try
  * {
- *     await waitForTheElement('.element-that-does-not-exist-yet', {
+ *     await waitForTheElement('.element-that-may-appear-later', {
  *         timeout : 5000
  *     });
  * }
  * catch (error)
  * {
- *     throw new Error('Took more than 5 seconds to fetch the element.');
+ *     throw new Error('Took more than 5 seconds to find the element.');
  * }
  * ```
  *
@@ -63,7 +63,7 @@ function getMatchingElementFromMutation  (mutation, selector)
  *
  * @param {String}  selector                   The selector of the element to fetch.
  * @param {Object}  [options]                  Some options to control how the element is searched for.
- * @param {Number}  [options.timeout = 2500]   Determines how long you want to wait for (in milliseconds) before an error is thrown.
+ * @param {Number}  [options.timeout = 2500]   Determines the maximum amount of time you want to wait for (in milliseconds).
  * @param {Element} [options.scope = document] Determines the scope you want to search in.
  *
  * @throws {Error} When a matching element isn't found in time.
@@ -120,4 +120,36 @@ function waitForTheElement (selector, {
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-module.exports = waitForTheElement;
+/**
+ * Fetches an element by waiting for it to exist.
+ *
+ * This works the same way as `waitForTheElement()`, but returns `null` instead of throwing an error.
+ *
+ * Example usage:
+ *
+ * ```
+ * let element = await tryWaitingForTheElement('.element-that-may-appear-later', {
+ *     timeout : 5000
+ * });
+ *
+ * if (element === null)
+ * {
+ *     console.log('Took more than 5 seconds to find the element.');
+ * }
+ * ```
+ *
+ * @returns {Promise<Element>} A promise that will be fulfilled with the matching element or `null` if a matching element could not be found.
+ *
+ * @param {String}  selector                   The selector of the element.
+ * @param {Object}  [options]                  Some options to control how the element is searched for.
+ * @param {Number}  [options.timeout = 2500]   Determines the maximum amount of time you want to wait for (in milliseconds).
+ * @param {Element} [options.scope = document] Determines the scope you want to search in.
+ */
+function tryAndWaitForTheElement (selector, options)
+{
+	return waitForTheElement(selector, options).catch(() => null);
+}
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+module.exports = { waitForTheElement, tryAndWaitForTheElement };