+++ tests/waitForTheElement.test.js
@@ -1,29 +1,31 @@
 'use strict';
 
+// Dependencies
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 const { expect, use } = require('chai');
 const chaiAsPromised  = require('chai-as-promised');
 const delay           = require('timeout-as-promise');
 
+// Subjects
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-const waitForTheElement = require('../src/waitForTheElement');
+const { waitForTheElement, tryAndWaitForTheElement } = require('../src/waitForTheElement');
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+before(function ()
+{
+	use(chaiAsPromised);
+});
+
+afterEach(function ()
+{
+	document.body.innerHTML = '';
+});
 
 describe('function waitForTheElement(selector, options)', function ()
 {
-	before(function ()
-	{
-		use(chaiAsPromised);
-	});
-
-	afterEach(function ()
-	{
-		document.body.innerHTML = '';
-	});
-
 	describe('when elements matching `selector` already exist', function ()
 	{
 		beforeEach(function ()
@@ -49,7 +51,7 @@ describe('function waitForTheElement(selector, options)', function ()
 			).to.eventually.have.property('id', 'target-1');
 		});
 
-		it('shall return a promise that will be fulfulled with the first matching element that is in `options.scope`', function ()
+		it('shall return a promise that will be fulfilled with the first matching element that is in `options.scope`', function ()
 		{
 			let scope = document.querySelector('#parent');
 
@@ -153,6 +155,157 @@ describe('function waitForTheElement(selector, options)', function ()
 
 			// Act.
 			let element = waitForTheElement('.target', {
+				scope
+			});
+
+			// Wait.
+			await delay(1000);
+
+			// Setup.
+			document.querySelector('#target-1').setAttribute('class', 'target');
+
+			// Setup.
+			scope.innerHTML = `
+				<div class="target" id="target-2"></div>
+			`;
+
+			// Assert.
+			return expect(element).to.eventually.have.property('id', 'target-2');
+		});
+	});
+});
+
+describe('function tryAndWaitForTheElement(selector, options)', function ()
+{
+	describe('when elements matching `selector` already exist', function ()
+	{
+		beforeEach(function ()
+		{
+			document.body.innerHTML = `
+				<div class="target" id="target-1"></div>
+
+				<div id="parent">
+					<div class="target" id="target-2">
+
+					</div>
+				</div>
+
+				<div class="target" id="target-3"></div>
+			`;
+		});
+
+		it('shall return a promise that will be fulfilled with the first matching element', function ()
+		{
+			// Act & Assert.
+			return expect(
+				tryAndWaitForTheElement('.target')
+			).to.eventually.have.property('id', 'target-1');
+		});
+
+		it('shall return a promise that will be fulfilled with the first matching element that is in `options.scope`', function ()
+		{
+			let scope = document.querySelector('#parent');
+
+			// Act & Assert.
+			return expect(
+				tryAndWaitForTheElement('.target', {
+					scope
+				})
+			).to.eventually.have.property('id', 'target-2');
+		});
+	});
+
+	describe('when no elements matching `selector` exist', function ()
+	{
+		it('shall return a promise that will be fulfilled with `null` when an element is not matched after 2.5 seconds', function ()
+		{
+			// Act & Assert.
+			return expect(
+				tryAndWaitForTheElement('.target')
+			).to.eventually.be.null;
+		});
+
+		it('shall return a promise that will be fulfilled with `null` when an element is not matched after `options.timeout`', function ()
+		{
+			// Act & Assert.
+			return expect(
+				tryAndWaitForTheElement('.target', {
+					timeout : 4000
+				})
+			).to.eventually.be.null;
+		});
+
+		it('shall return a promise that will be fulfilled with a matching element when it is eventually added', async function ()
+		{
+			// Act.
+			let element = tryAndWaitForTheElement('.target');
+
+			// Wait.
+			await delay(1000);
+
+			// Setup.
+			document.body.innerHTML = `
+				<div class="target" id="target-1"></div>
+				<div class="target" id="target-2"></div>
+			`;
+
+			// Assert.
+			return expect(element).to.eventually.have.property('id', 'target-1');
+		});
+
+		it('shall return a promise that will be fulfilled with a matching element, even when it is added as a child of another added element', async function ()
+		{
+			// Act.
+			let element = tryAndWaitForTheElement('.target');
+
+			// Wait.
+			await delay(1000);
+
+			// Setup.
+			document.body.innerHTML = `
+				<div id="parent">
+					<div class="target" id="target"></div>
+				</div>
+			`;
+
+			// Assert.
+			return expect(element).to.eventually.have.property('id', 'target');
+		});
+
+		it('shall return a promise that will be fulfilled with a matching element after an attribute change', async function ()
+		{
+			// Setup.
+			document.body.innerHTML = `
+				<div id="target"></div>
+			`;
+
+			// Act.
+			let element = tryAndWaitForTheElement('.target');
+
+			// Wait.
+			await delay(1000);
+
+			// Setup.
+			document.querySelector('#target').setAttribute('class', 'target');
+
+			// Assert.
+			return expect(element).to.eventually.have.property('id', 'target');
+		});
+
+		it('shall return a promise that will be fulfilled with a matching element only when it is in `options.scope`', async function ()
+		{
+			// Setup.
+			document.body.innerHTML = `
+				<div id="target-1"></div>
+				<div id="parent">
+
+				</div>
+			`;
+
+			let scope = document.querySelector('#parent');
+
+			// Act.
+			let element = tryAndWaitForTheElement('.target', {
 				scope
 			});