+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] (2018-09-25)
+
+### Changed
+
+  - The module no longer exports a function.
+
+### Added
+
+  - Introduced the `tryAndWaitForTheElement()` function that will return `null` instead of throwing an error when a matching element is not found.
+
 ## [1.0.0] (2018-09-25)
 
 The initial public release.